Move isBlockAllowedOnPage function inline
Remove last remaining usage of that function as all other components use the new `useIsBlockAllowedOnPage` hook.
