+++ assets/src/stories-editor/helpers/index.js
@@ -60,5 +60,4 @@ export { default as parseDropEvent } from './parseDropEvent';
 export { default as getRelativeElementPosition } from './getRelativeElementPosition';
 export { default as isCTABlock } from './isCTABlock';
 export { default as useMoveBlockToPage } from './useMoveBlockToPage';
-export { default as isBlockAllowedOnPage } from './isBlockAllowedOnPage';
 export { default as useIsBlockAllowedOnPage } from './useIsBlockAllowedOnPage';
